+++ noodles-cram/examples/cram_view_async.rs
@@ -4,7 +4,7 @@ use futures::TryStreamExt;
 use noodles_cram as cram;
 use noodles_fasta::{self as fasta, repository::adapters::IndexedReader};
 use noodles_sam as sam;
-use tokio::{fs::File, io};
+use tokio::io;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -20,12 +20,16 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .map(fasta::Repository::new)
         .unwrap_or_default();
 
-    let mut reader = File::open(src).await.map(cram::AsyncReader::new)?;
+    let mut reader = cram::r#async::io::reader::Builder::default()
+        .set_reference_sequence_repository(repository)
+        .build_from_path(src)
+        .await?;
+
     reader.read_file_definition().await?;
 
     let header = reader.read_file_header().await?.parse()?;
 
-    let mut records = reader.records(&repository, &header);
+    let mut records = reader.records(&header);
 
     let mut writer = sam::AsyncWriter::new(io::stdout());